requirements.txt
psycopg[binary]==3.2.6
python-dotenv==1.0.1
